# app/backend/contracts/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/vision/rollout": {
      "get": {
        "operationId": "RolloutController_rollout",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Global + per-modality + per-provider rollout stages, with reasons.",
        "tags": [
          "vision-rollout"
        ]
      }
    },
    "/api/vision/trust": {
      "get": {
        "operationId": "RolloutController_trust",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Derived trust analytics — user/provider/modality/food/portion.",
        "tags": [
          "vision-rollout"
        ]
      }
    },
...
    "/api/vision/risk": {
      "get": {
        "operationId": "RolloutController_risk",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Five-dimension deployment risk, every claim with its number.",
        "tags": [
          "vision-rollout"
        ]
      }
    },
    "/api/vision/timeline": {
      "get": {
        "operationId": "RolloutController_timeline",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the identity is not on the ADMIN_EMAILS allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Weekly trust evolution from append-only history.",
        "tags": [
          "vision-rollout"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vitals Fit API",
    "description": "Nutrition platform API. Endpoints under /vision/{rollout,governance,promotion,rollback,canary} and the platform-wide /vision/learning analytics are OPERATOR-only (ADMIN_EMAILS allowlist).",
    "version": "5.5",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "InternalErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 500
          },
          "message": {
            "type": "string",
            "example": "Internal server error",
            "description": "Always this literal string — never the real cause."
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-22T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/vision/scans",
            "description": "The route that failed, for log correlation."
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "HttpErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "Repeats the HTTP status code."
          },
          "message": {
            "description": "A human-readable message, or one string per failed validation rule.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Resource not found."
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Nest's standard reason phrase for the status code."
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
    }
  }
}